Change tuntap_set_ip() interface
The third parameter, the netmask, is now an integer.
It means that instead of sending "255.255.255.0", you send 24.

+++ tuntap-linux.c
@@ -114,7 +114,7 @@ tuntap_sys_set_hwaddr(struct device *dev, struct ether_addr *eth_addr) {
 }
 
 int
-tuntap_sys_set_ip(struct device *dev, unsigned int iaddr, unsigned int imask) {
+tuntap_sys_set_ip(struct device *dev, unsigned int iaddr, unsigned long imask) {
 	struct sockaddr_in addr;
 	struct ifreq ifr;
